refactor: polyfill components
