+++ lib/java/polyglot-ydoc-server/src/main/java/module-info.java
@@ -1,5 +1,4 @@
-
-module PolyglotYdocServer {
+module YdocServer {
   requires org.graalvm.polyglot;
   requires io.helidon.webclient;
   requires io.helidon.webclient.websocket;
