+++ lib/java/polyglot-ydoc-server/src/main/java/org/enso/ydoc/Polyfill.java
@@ -1,8 +1,8 @@
-package org.enso.polyfill;
+package org.enso.ydoc;
 
 import org.graalvm.polyglot.Context;
 
 public interface Polyfill {
 
-    void initialize(Context ctx);
+  void initialize(Context ctx);
 }
