+++ lib/java/polyglot-ydoc-server/src/main/resources/org/enso/ydoc/polyfill/encoding.js
