+++ lib/java/polyglot-ydoc-server/src/main/resources/org/enso/ydoc/polyfill/websocket.js
